+++ src/instance.rs
@@ -11,9 +11,6 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::{bail, Context, Result};
-use ra_multiplex::config::Config;
-use ra_multiplex::lsp::{self, Message};
-use ra_multiplex::proto;
 use serde_json::{Number, Value};
 use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::process::{Child, ChildStderr, ChildStdin, ChildStdout, Command};
@@ -21,7 +18,10 @@ use tokio::sync::{mpsc, Mutex, Notify, RwLock};
 use tokio::{select, task, time};
 use tracing::{debug, error, info, info_span, instrument, trace, warn, Instrument};
 
-use super::async_once_cell::AsyncOnceCell;
+use crate::async_once_cell::AsyncOnceCell;
+use crate::config::Config;
+use crate::lsp::{self, Message};
+use crate::proto;
 
 /// keeps track of the initialize/initialized handshake for an instance
 #[derive(Default)]
