+++ src/lib.rs
@@ -1,3 +1,9 @@
-pub mod config;
-pub mod lsp;
-pub mod proto;
+mod async_once_cell;
+mod client;
+mod config;
+mod instance;
+mod lsp;
+mod proto;
+
+pub mod proxy;
+pub mod server;
